Fix min function deduction on ESP platforms
Having the calculation as part of the function argument was apparently confusing the template. Casting the result of the uint16 - uint16 math to a uint16 fixes the problem and it can find the template just fine.

This wasn't previously an issue on the AVR platforms because the min function is done using a macro there.